[SPARK-41993][SQL] Move RowEncoder to AgnosticEncoders
This PR makes `RowEncoder` produce an `AgnosticEncoder`. The expression generation for these encoders is moved to `ScalaReflection` (this will be moved out in a subsequent PR).

The generated serializer and deserializer expressions will slightly change for both schema and type based encoders. These are not semantically different from the old expressions. Concretely the following changes have been introduced:
- There is more type validation in maps/arrays/seqs for type based encoders. This should be a positive change, since it disallows users to pass wrong data through erasure hackd.
- Array/Seq serialization is a bit more strict. In the old scenario it was possible to pass in sequences/arrays with the wrong type and/or nullability.

For the Spark Connect Scala Client we also want to be able to use `Row` based results.

No

This is a refactoring, existing tests should be sufficient.

Closes #39517 from hvanhovell/SPARK-41993.

Authored-by: Herman van Hovell <herman@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>